working on myglibc

+++ kernel/include/linux/kernel/types.h
@@ -28,7 +28,7 @@
 		typedef __kernel_dev_t		dev_t;
 		typedef __kernel_ulong_t	ino_t;
 		typedef __kernel_mode_t		mode_t;
-		typedef unsigned short		umode_t;
+		typedef unsigned int		umode_t;
 		typedef u32					nlink_t;
 		typedef __kernel_off_t		off_t;
 		typedef __kernel_pid_t		pid_t;
